+++ jee-mongodb-backend/geo-service/src/main/java/de/zeiss/mongodbws/geoservice/persistence/entity/GeoPoint.java
