+++ jee-mongodb-backend/geo-service/src/main/java/de/zeiss/mongodbws/geoservice/rest/Constants.java
@@ -1,0 +1,22 @@
+/**
+ * This file is part of a demo application showing MongoDB usage with Morphia library.
+ * 
+ * Copyright (C) 2016 Saxonia Systems AG
+ */
+package de.saxsys.mongodbws.geoservice.rest;
+
+/**
+ * 
+ * @author Andreas Post
+ */
+public abstract class Constants {
+
+	public static final String CONTENT_ENC_KEY = "Content-Encoding";
+
+	public static final String CHARSET_UTF8 = "UTF-8";
+
+	public static final String MEDIA_TYPE_JSON = "application/json";
+
+	public static final String POI_RESOURCE_PATH = "rest/poi/";
+
+}
