+++ jee-mongodb-backend/geo-service/src/test/java/de/zeiss/mongodbws/geoservice/rest/resource/PointOfInterestResourceControllerTest.java
