+++ jee-mongodb-backend/geo-service/src/test/java/de/zeiss/mongodbws/geoservice/service/PointOfInterestEntityConverterTest.java
